Bug 16900 fixed: no-gui build of Scilab raised an error at startup
call to `helpbrowser_update()` macro was made in `etc/scilab.start` regardless of scilab launch mode. This is OK if scilab has been built with gui since the macro is built, but not OK if Scilab is built without Java, which enforces the use of `--disable-build-help` and cancels the loading of `helptools` module (which is partly built without Java stuff).

Fix #16900

Closes #16900
See https://gitlab.com/scilab/scilab/-/merge_requests/25