atoms: atomsInstall will execute builder.sce if no binary version is available
`atomsInstall()` was not using the `source` prefixed content to download and install the source version of a toolbox. With this change, it will download and try to build sources in the same local directory as before. Binary will be preferred if available.

Co-authored-by: Clement DAVID <clement.david@3ds.com>
Closes #13902
See https://gitlab.com/scilab/scilab/-/merge_requests/173