fix #13087: update macOS link vs. JNI and Java detection in scilab script
Fixes legacy issue #13087 of Scilab binary dependency on JavaVM.framework which required installing a system JRE/JDK even if a built-in jre was provided. After this patch, the macOS version of Scilab can be released without or with an integrated JRE and in the latter case, the user will no longer be prompted to install one.

Co-authored-by: Clément DAVID <clement.david@3ds.com>
Closes #13087
See https://gitlab.com/scilab/scilab/-/merge_requests/207