make test_run log files into the JUnit XML
`test_run` was not logging stdout and stderr to the resulting XML file. Errors were hard to find without access to the files.

This commit also distinguishes failures (in the test script) and errors (where the Scilab exit code is not 0) in the resulting JUnit XML report.

See https://gitlab.com/scilab/scilab/-/merge_requests/448