CI: build prerequirements when needed on Linux
When updating a dependency, we previously had to update requirements first and a Merge Request later (breaking other merge requests). This commit builds the pre-requirements in the same MR and uses it for the current build. This will keep all the needed modifications in the same MR and allow iterating without breaking other MRs.

This commit also:
* Moves and update the Linux script used to build prereq from SVN to Git.
* Updates the Dockerfile.linux and build job according to the new prereq package.
* Creates a Dockerfile used to build prereqs.
* Updates all download links to use `Outscale` for prereq sources.
* Removes unneeded content: `docbook` and manpages

/!\ jar files located in the `thirdparty` folder have been copied from the actual prereqs /!\

Co-authored-by: Clément DAVID <clement.david@3ds.com>
See https://gitlab.com/scilab/scilab/-/merge_requests/294