Fix macOS arm64 startup
When Scilab is started from the app bundle the architecture is always set to x86_64 by Rosetta as bin/scilab is a script shell. Architecture has to be detected on scilab-bin and forced. This is not necessary for the development version as bin/scilab is launched from the command line.

See https://gitlab.com/scilab/scilab/-/merge_requests/531