fix #16935 - Variable was modified after a call to a graphics function
As for Scilab 5, set the module graphics in copy mode by removing the `intppty("graphics")`. (reverting b49732bb49035758a3d6469eceafd3a13d113e6e)<br>
That fixes the issue but this also makes crash the `demo_gui` because of a double free.

demo_gui.sce:
```
demo_fig.userdata = struct();
demo_fig.userdata.frame_number = 1;
```

In the second line, at the insertion of `userdata` in the graphic handle, the `struct` to be inserted is clone due to the copy mode.<br>
Because this `struct` is only contained in the graphic handle, it will be deleted before setting the cloned one. (see set_user_data_property.c, deleting the property before setting the new one)<br>
The crash appends because the original struct is also contained in an `ExpHistory` object due to the extraction of `demo_fig.userdata` and will be deleted a second time later.

It worked in the ref mode because we are setting in the graphic handle the same object that is protected in `set_user_data_property.c`

Closes #16935
See https://gitlab.com/scilab/scilab/-/merge_requests/208