Fix #16950 - `tree_show()` improved display of regular arrays
`tree_show()` display of regular array elements was not optimal: Either the size was not displayed while it was not trivial to see from displayed values or no value was shown. In contrast, the limited series of first ones could be helpful.

Co-authored-by: Clement DAVID <clement.david@3ds.com>
Closes #16950
See https://gitlab.com/scilab/scilab/-/merge_requests/370