CI: introduce a build tag for runners
The Scilab tests are slow enough to lock CI and not load the runner's CPU. This MR introduces a dedicated tag for `build` jobs that use internal parallelism to busy the CPU; the number of CI parallel jobs is limited in the associated runner. Other jobs are not executing processes in parallel; the number of CI parallel jobs is high.

See https://gitlab.com/scilab/scilab/-/merge_requests/591