Scinotes backup files were not always deleted
fixes #17108 partly. Besides the fix of the main issue, the proposal improves the autosave/backup feature by deleting backup files when the editor pane is closed (and not when Scilab exits). This is the actual behavior of most text editors.

This MR does not fix the following sub-issues:
1. when a file is saved in a different folder than the current directory, the backup file is always created in the current directory.
2. when editing a new file from scratch, there is no backup creation until the current file is saved by the user.

Closes #17108
See https://gitlab.com/scilab/scilab/-/merge_requests/553