Use SSPInputConnector and SSPOutputConnector
This blocks are used on the main diagram connectors and contains an
implementation that vary depending on some conditions. If there is an
SSPInputConnector in a schema, we are in a "function call mode" : inputs
are mapped to FROMWSB and outputs are mapped to TOWS_c. If not, outputs
are mapped to CSCOPE blocks.