fix #17185 - callback execution at exit and -nogui argument fixed
The freeze was due to an extra execution of a priority command (a graphic callback) when Scilab was in exit mode (force quit == true).
Because of the exit mode, the tread that unstacks commands was closed, so the callback could not be executed and remained in the store command.
The execution of `scilab.quit` gets interrupted due to the priority of the callback command that makes Scilab wait for a signal that must be sent by the closed thread.
This fix check that the command store is well empty before leaving the thread.


Fix `-nogui` to have the same behavior than `-nwni`.

Closes #17185
See https://gitlab.com/scilab/scilab/-/merge_requests/726